Moving receiver and background to base and using Time on dates
