+++ radiospectra/spectrogram/sources/swaves.py
@@ -7,10 +7,7 @@ __all__ = ["SWAVESSpectrogram", "SWAVESMeta"]
 class SWAVESMeta(SpectrogramMeta):
     """Metadata for STEREO/SWAVES spectrograms."""
 
-    @property
-    def receiver(self) -> str | None:
-        """The name of the receiver."""
-        return self.get("detector")
+    pass
 
 
 class SWAVESSpectrogram(GenericSpectrogram):
@@ -37,11 +34,6 @@ class SWAVESSpectrogram(GenericSpectrogram):
         if not isinstance(meta, SWAVESMeta):
             meta = SWAVESMeta(meta)
         super().__init__(meta=meta, data=data, **kwargs)
-
-    @property
-    def receiver(self):
-        """The name of the receiver."""
-        return self.meta.receiver
 
     @classmethod
     def is_datasource_for(cls, data, meta, **kwargs):
