+++ radiospectra/spectrogram/tests/test_meta.py
@@ -1,5 +1,10 @@
 import pytest
 
+import astropy.units as u
+from astropy.time import Time
+
+from sunpy.net import attrs as a
+
 from radiospectra.spectrogram.meta import SpectrogramMeta
 
 
@@ -17,8 +22,8 @@ def test_spectrogram_meta_required_keys():
     assert meta.instrument == "test_inst"
     assert meta.observatory == "test_obs"
     assert meta.detector == "test_det"
-    assert meta.date_start == "2020-01-01"
-    assert meta.date_end == "2020-01-02"
+    assert meta.date_start == Time("2020-01-01")
+    assert meta.date_end == Time("2020-01-02")
 
 
 def test_spectrogram_meta_missing_required():
@@ -32,11 +37,11 @@ def test_spectrogram_meta_optional_keys():
         {
             "processing_level": "L2",
             "version": "1.0",
-            "wavelength": "radio",
+            "wavelength": a.Wavelength(1 * u.kHz, 10 * u.kHz),
         }
     )
 
     assert meta.processing_level == "L2"
     assert meta.version == "1.0"
-    assert meta.frequency_range == "radio"
+    assert meta.frequency_range == a.Wavelength(1 * u.kHz, 10 * u.kHz)
     assert meta.temporal_resolution is None
